requirements.txt
algoliasearch==1.17.0
algoliasearch-django==1.7.0
astroid==2.0.4
autopep8==1.4.1
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
cloudinary==1.12.0
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.1.1
django-allauth==0.38.0
django-cloudinary-storage==0.2.3
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
oauthlib==2.1.0
pbr==4.2.0
Pillow==5.2.0
pycodestyle==2.4.0
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
whitenoise==3.3.1
wrapt==1.10.11
psycopg2==2.7.3.1
